backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.0

# ML Libraries (using versions with pre-built Windows wheels)
scikit-learn==1.5.2
xgboost==2.1.3
shap==0.46.0
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.92.1

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dotenv==1.0.0
